astype, set make copy to false

+++ model.py
@@ -35,7 +35,7 @@ class AgenderNetVGG16(Model):
 
 	@staticmethod
 	def prepImg(data):
-		data = data.astype('float16')
+		data = data.astype('float16', copy=False)
 		data = data[..., ::-1]
 		mean = [103.939, 116.779, 123.68]
 		data[..., 0] -= mean[0]
@@ -67,7 +67,7 @@ class AgenderNetInceptionV3(Model):
 
 	@staticmethod
 	def prepImg(data):
-		data = data.astype('float16')
+		data = data.astype('float16', copy=False)
 		data /= 127.5
 		data -= 1.
 		return data
@@ -96,7 +96,7 @@ class AgenderNetXception(Model):
 
 	@staticmethod
 	def prepImg(data):
-		data = data.astype('float16')
+		data = data.astype('float16', copy=False)
 		data /= 127.5
 		data -= 1.
 		return data
@@ -128,7 +128,7 @@ class AgenderNetMobileNetV2(Model):
 	def prepImg(data):
 		data = [cv2.resize(image, (96,96), interpolation = cv2.INTER_CUBIC) for image in data]
 		data = np.array(data)
-		data = data.astype('float32')
+		data = data.astype('float16', copy=False)
 		data /= 128.
 		data -= 1.
 		return data
